+++ src/main/java/algorithms/TileBasedMap.java
@@ -43,7 +43,7 @@ public interface TileBasedMap {
 	 * @param y The y coordinate of the tile to check
 	 * @return True if the location is blocked
 	 */
-	public boolean blocked(Mover mover, int x, int y);
+	public boolean blocked(Mover mover, double x, double y);
 	
 	/**
 	 * Get the cost of moving through the given tile. This can be used to
